changes to main program
changed the way the LCD updates instead of every 9-10 seconds. it now changes along with the sensor readings.
added test code for the fans. i can setup a temp as a limit and above it they will turn on, below it they will turn off.

autorun main programm at boot of Raspberry pi.